remove captcha feature flag
