+++ app/api/auth/index.js
@@ -1,5 +1,3 @@
 import needsAuthorization from './authMiddleware.js';
-import captchaAuthorization from './captchaMiddleware.js';
-import { CaptchaModel } from './CaptchaModel.js';
 
-export { needsAuthorization, captchaAuthorization, CaptchaModel };
+export { needsAuthorization };
