+++ app/api/auth/publicAPIMiddleware.ts
@@ -1,6 +1,6 @@
 import type { Request, Response, NextFunction } from 'express';
 import settings from '#api/settings/index.js';
-import { captchaAuthorization } from './index.js';
+import { captchaMiddleware } from '#api/core/infrastructure/express/captcha/CaptchaMiddleware.js';
 
 export const publicAPIMiddleware = async (req: Request, res: Response, next: NextFunction) => {
   const { openPublicEndpoint } = await settings.get();
@@ -10,5 +10,5 @@ export const publicAPIMiddleware = async (req: Request, res: Response, next: Nex
     return next();
   }
 
-  return captchaAuthorization()(req, res, next);
+  return captchaMiddleware()(req, res, next);
 };
