+++ app/api/auth/routes.js
@@ -2,20 +2,11 @@ import cookieParser from 'cookie-parser';
 import MongoStore from 'connect-mongo';
 import passport from 'passport';
 import session from 'express-session';
-import svgCaptcha from 'svg-captcha';
-import urljoin from 'url-join';
-import cors from 'cors';
-import settings from '#api/settings/index.js';
 import { DB } from '#api/odm/index.js';
 import { config } from '#api/config.js';
-import request from '#shared/JSONRequest.js';
-import { tenants } from '#api/tenants/index.js';
 import { LoginController } from '#api/core/infrastructure/express/users/LoginController.js';
 import { LogoutController } from '#api/core/infrastructure/express/users/LogoutController.js';
 import { GetCurrentUserController } from '#api/core/infrastructure/express/users/GetCurrentUserController.js';
-import { CaptchaController } from '#api/core/infrastructure/express/captcha/CaptchaController.js';
-import { RemoteCaptchaController } from '#api/core/infrastructure/express/captcha/RemoteCaptchaController.js';
-import { CaptchaModel } from './CaptchaModel.js';
 
 import './passport_conf.js';
 
@@ -50,41 +41,6 @@ const authRoutes = app => {
   app.get('/api/user', GetCurrentUserController.createHandler());
 
   app.get('/logout', LogoutController.createHandler());
-
-  const corsOptions = {
-    origin: true,
-    methods: 'GET',
-    credentials: true,
-    optionsSuccessStatus: 200,
-  };
-
-  app.get('/api/captcha', cors(corsOptions), async (req, res) => {
-    if (tenants.current().featureFlags?.v2Captcha) {
-      await CaptchaController.createHandler()(req, res);
-      return;
-    }
-
-    // @deprecated v1 fallback for the `v2Captcha` flag, superseded by CaptchaController.
-    // Remove once v2Captcha is enabled for all tenants.
-    const captcha = svgCaptcha.create({ ignoreChars: '0OoiILluvUV' });
-    const text = process.env.DATABASE_NAME !== 'uwazi_e2e' ? captcha.text : '42hf';
-    const storedCaptcha = await CaptchaModel.save({ text });
-
-    res.json({ svg: captcha.data, id: storedCaptcha._id.toString() });
-  });
-
-  app.get('/api/remotecaptcha', async (req, res) => {
-    if (tenants.current().featureFlags?.v2Captcha) {
-      await RemoteCaptchaController.createHandler()(req, res);
-      return;
-    }
-
-    // @deprecated v1 fallback for the `v2Captcha` flag, superseded by RemoteCaptchaController.
-    // Remove once v2Captcha is enabled for all tenants.
-    const { publicFormDestination } = await settings.get({}, { publicFormDestination: 1 });
-    const remoteResponse = await request.get(urljoin(publicFormDestination, '/api/captcha'));
-    res.json(remoteResponse.json);
-  });
 };
 
 export { populateAuthenticatedUser, authenticatedUserMiddlewares };
