+++ app/api/config.ts
@@ -165,6 +165,8 @@ export const config = {
       aiAssistant: FEATURE_FLAG_AI_ASSISTANT === 'true' || false,
       aiAssistantServiceUrl: AI_ASSISTANT_SERVICE_URL || undefined,
       usersDirectory: false,
+      v2PrivateInstance: false,
+      v2Usergroups: false,
       telemetry: {
         enabled: false,
         sampleRate: 0.5,
@@ -173,9 +175,6 @@ export const config = {
         enabled: false,
         sampleRate: 0.5,
       },
-      v2Captcha: false,
-      v2PrivateInstance: false,
-      v2Usergroups: false,
     },
   },
   externalServices: (process.env.EXTERNAL_SERVICES || '').toLowerCase() === 'true',
