+++ app/api/contact/routes.js
@@ -1,11 +1,11 @@
 import contact from './contact.js';
-import { captchaAuthorization } from '../auth/index.js';
+import { captchaMiddleware } from '../core/infrastructure/express/captcha/CaptchaMiddleware.js';
 import { validation } from '../utils/index.js';
 
 export default app => {
   app.post(
     '/api/contact',
-    captchaAuthorization(),
+    captchaMiddleware(),
     validation.validateRequest({
       type: 'object',
       properties: {
