+++ app/api/contact/specs/routes.spec.js
@@ -8,9 +8,11 @@ jest.mock('../../utils/languageMiddleware.ts', () => (_req, _res, next) => {
   next();
 });
 
-jest.mock('../../auth/captchaMiddleware.ts', () => () => (_req, _res, next) => {
-  next();
-});
+jest.mock('../../core/infrastructure/express/captcha/CaptchaMiddleware.ts', () => ({
+  captchaMiddleware: () => (_req, _res, next) => {
+    next();
+  },
+}));
 
 describe('contact', () => {
   const app = setUpApp(contactRoutes);
