+++ app/api/tenants/tenantsModel.ts
@@ -59,7 +59,6 @@ const mongoSchema = new mongoose.Schema({
       enabled: Boolean,
       sampleRate: Number,
     },
-    v2Captcha: Boolean,
     v2PrivateInstance: Boolean,
     v2Usergroups: Boolean,
   },
