+++ app/server.js
@@ -23,6 +23,7 @@ import uwaziMessage from '../message.js';
 import apiRoutes from './api/api.js';
 import privateInstanceMiddleware from './api/auth/privateInstanceMiddleware.js';
 import authRoutes, { populateAuthenticatedUser } from './api/auth/routes.js';
+import { captchaRoutes } from '#api/core/infrastructure/express/captcha/routes.js';
 import { config } from './api/config.js';
 
 import { versionRoutes } from './api/version/routes.js';
@@ -123,6 +124,7 @@ DB.connect(config.DBHOST, config.DBAUTH).then(async () => {
   populateAuthenticatedUser(app);
   app.use(dependenciesContextMiddleware);
   authRoutes(app);
+  captchaRoutes(app);
   versionRoutes(app);
   app.use(privateInstanceMiddleware);
   app.use('/flag-images', express.static(path.resolve(__dirname, '../dist/flags')));
